Add support for arbitrary user keywords
(deftag foo/bar ..) will now define the functions ->bar, bar? and the var bar,
all of which operate on tagged values with the tag :foo/bar. This allows users
to create and use arbitrary symbolic keys rather than enforcing namespacing. If
no namespace is provided, then *ns* will be used as the source of the tag's
namespace same as before.