Differentiate safe and unsafe reader functions
read-tagged-val that was used for reading #g/t and #guten/tag literals
called eval on its argument. If I want to read something with
clojure.edn/read(-string), I cannot use this function, because it
violates clojure.edn's safety guarantees.

Move the evaluating behaviour to unsafe-read-tagged-val and thereby make
read-tagged-val safe.

The data readers in data_readers.clj continue pointing to
read-tagged-val, which now is the safe flavour of the reader function as
opposed to the unsafe one before. I consider this a better default, so
don't update them to point to unsafe-read-tagged-val.